#modified:   README.md #modified:   index.js #modified:   package.json
Modified index.js to add an optional argument to every method dealing with Routes and Targets.
This argument allows inlining an Options {Object} that overrides default schemes, hosts, and ports.

The README.md reflects that change.
The package.json bumps the version number to reflect backward compatible changes with a MINOR bump.